@@ -15,31 +15,58 @@
 package augeas
 
 import (
-	"io/ioutil"
+	"embed"
+	"fmt"
+	"io/fs"
+	"os"
 	"path/filepath"
-
-	"github.com/pkg/errors"
 )
 
-//go:generate go-bindata -nometadata -nocompress -pkg augeas -o assets.go assets/lenses
+//go:embed assets/*
+var assets embed.FS
+
+const lensSubdir = "assets/lenses"
 
 func InstallLenses(targetDir string) error {
-	lenses, err := AssetDir("assets/lenses")
+	lensesFS, err := fs.Sub(assets, lensSubdir)
 	if err != nil {
-		return errors.Wrap(err, "listing lenses")
+		return fmt.Errorf("reading lenses subdirectory %s: %w", lensSubdir, err)
 	}
 
-	for _, lens := range lenses {
-		lensBytes, err := Asset("assets/lenses/" + lens)
-		if err != nil {
-			return errors.Wrapf(err, "fetching lens %s", lens)
-		}
-
-		err = ioutil.WriteFile(filepath.Join(targetDir, lens), lensBytes, 0644)
-		if err != nil {
-			return errors.Wrapf(err, "writing lens %s", lens)
-		}
-
+	if err := fs.WalkDir(lensesFS, ".", genCopyLensToDiskFunc(lensesFS, targetDir)); err != nil {
+		return fmt.Errorf("walking lenes directory: %w", err)
 	}
+
 	return nil
+}
+
+// genCopyLensToDiskFunc returns fs.WalkDirFunc function that will
+// copy files to disk in a given location.
+func genCopyLensToDiskFunc(srcFS fs.FS, targetdir string) fs.WalkDirFunc {
+	return func(path string, d fs.DirEntry, err error) error {
+		if err != nil {
+			// Should be impossible for embed https://github.com/golang/go/issues/45815
+			return fmt.Errorf("Impossible error from embed: %w", err)
+		}
+
+		if path == lensSubdir || path == "." {
+			return nil
+		}
+
+		if d.IsDir() {
+			return fmt.Errorf("Lenses can only be files, %s is a directory", path)
+		}
+
+		data, err := fs.ReadFile(srcFS, path)
+		if err != nil {
+			// Should be impossible for embded https://github.com/golang/go/issues/45815
+			return fmt.Errorf("Impossible error from embed: %w", err)
+		}
+
+		if err := os.WriteFile(filepath.Join(targetdir, path), data, 0644); err != nil {
+			return fmt.Errorf("writing lens %s, got: %w", path, err)
+		}
+
+		return nil
+	}
 }
