@@ -3,24 +3,26 @@ package packaging
 import (
 	"bytes"
 	"context"
+	"embed"
 	"encoding/json"
 	"fmt"
 	"io"
 	"io/ioutil"
 	"os"
 	"os/exec"
+	"path"
 	"path/filepath"
 	"strings"
 	"text/template"
 
 	"github.com/kolide/kit/fs"
 	"github.com/kolide/launcher/pkg/packagekit"
-	"github.com/kolide/launcher/pkg/packaging/internal"
 	"github.com/pkg/errors"
 	"go.opencensus.io/trace"
 )
 
-//go:generate go-bindata -nometadata -nocompress -pkg internal -o internal/assets.go internal/assets/
+//go:embed assets/*
+var assets embed.FS
 
 const (
 	// Enroll secret should be readable only by root
@@ -453,9 +455,9 @@ func (p *PackageOptions) renderLogrotateConfig(ctx context.Context) error {
 		PidPath: filepath.Join(p.rootDir, "launcher.pid"),
 	}
 
-	logrotateTemplate, err := internal.Asset("internal/assets/logrotate.conf")
+	logrotateTemplate, err := assets.ReadFile("assets/logrotate.conf")
 	if err != nil {
-		return errors.Wrapf(err, "failed to get template named %s", "internal/assets/logrotate.conf")
+		return errors.Wrapf(err, "failed to get template named %s", "assets/logrotate.conf")
 	}
 
 	tmpl, err := template.New("logrotate").Parse(string(logrotateTemplate))
@@ -598,7 +600,7 @@ func (p *PackageOptions) setupPreinst(ctx context.Context) error {
 		return nil
 	}
 
-	preinstallContent, err := internal.Asset("internal/assets/preinstall-darwin.sh")
+	preinstallContent, err := assets.ReadFile("assets/preinstall-darwin.sh")
 	if err != nil {
 		return errors.Wrap(err, "getting template for preinstall")
 	}
@@ -618,13 +620,13 @@ func (p *PackageOptions) setupPostinst(ctx context.Context) error {
 
 	switch {
 	case p.target.Platform == Darwin && p.target.Init == LaunchD:
-		postinstTemplateName = "internal/assets/postinstall-launchd.sh"
+		postinstTemplateName = "postinstall-launchd.sh"
 	case p.target.Platform == Linux && p.target.Init == Systemd:
-		postinstTemplateName = "internal/assets/postinstall-systemd.sh"
+		postinstTemplateName = "postinstall-systemd.sh"
 	case p.target.Platform == Linux && (p.target.Init == Upstart || p.target.Init == UpstartAmazonAMI):
-		postinstTemplateName = "internal/assets/postinstall-upstart.sh"
+		postinstTemplateName = "postinstall-upstart.sh"
 	case p.target.Platform == Linux && p.target.Init == Init:
-		postinstTemplateName = "internal/assets/postinstall-init.sh"
+		postinstTemplateName = "postinstall-init.sh"
 	default:
 		// If we don't match in the case statement, log that we're ignoring
 		// the setup, and move on. Don't throw an error.
@@ -632,7 +634,7 @@ func (p *PackageOptions) setupPostinst(ctx context.Context) error {
 		return nil
 	}
 
-	postinstTemplate, err := internal.Asset(postinstTemplateName)
+	postinstTemplate, err := assets.ReadFile(path.Join("assets", postinstTemplateName))
 	if err != nil {
 		return errors.Wrapf(err, "Failed to get template named %s", postinstTemplateName)
 	}
